% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_game_feed.R
\name{fetch_game_feed}
\alias{fetch_game_feed}
\title{fetch_game_feed}
\usage{
fetch_game_feed(game_id)
}
\arguments{
\item{game_id}{a number or string that represent the id of an nhl game}
}
\value{
data.table that contains all event for the requested game
}
\description{
fetch the game feed for a specific game from the nhl api
}
