% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_player_info.R
\name{fetch_player_info}
\alias{fetch_player_info}
\title{fetch_player_info}
\usage{
fetch_player_info(player_id)
}
\arguments{
\item{player_id}{a number or string that represent the id of an nhl player}
}
\value{
data.table that contains all info for the requested player
}
\description{
fetch miscellaneous info about specific player from the nhl api
}
\examples{
fetch_player_info(8470642)
fetch_player_info("8470642")
}
