% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_team_roster.R
\name{fetch_team_roster}
\alias{fetch_team_roster}
\title{fetch_team_roster}
\usage{
fetch_team_roster(team_id)
}
\arguments{
\item{team_id}{a number or string that represent the id of an nhl team}
}
\value{
data.table that contains data for an nhl team roster
}
\description{
fetch all team roster data from the nhl api
}
