% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_teams.R
\name{fetch_teams}
\alias{fetch_teams}
\title{fetch_teams}
\usage{
fetch_teams()
}
\value{
data.table that contains data for all nhl teams
}
\description{
fetch all teams data from the nhl api
}
\examples{
fetch_teams()
}
